HBSD: Fix build after upstream r351729
FreeBSD svn revision 351729 (git: 286ae5bd6b5d86148741a589f277ac299e3cff4e)
migrated sysctlbyname(3) to being a syscall. r351729 caused fallout in
HardenedBSD, due to HardenedBSD using sysctlbyname(3) in the RTLD. No
longer is sysctlbyname(3) a symbol in libc in the traditional sense. We
have to tell the build framework where to find the include files that
the stub sysctlbyname.c file requires and to generate the syscall stub
for __sysctlbyname.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>